+++ src/imio/history/interfaces.py
@@ -36,7 +36,8 @@ class IImioWfHistory(IImioHistory):
         """Returns the WF history for context."""
 
     def historyLastEventHasComments(self):
-        """Returns True if the last event of the object's history has a comment."""
+        """Returns True if the last event of the object's history has a comment.
+           Ideally, this method should return a list of unicode so comparison with comments is possible."""
 
     def ignorableHistoryComments(self):
         """Ignorable history comments, stored as utf-8."""